Fix fields losing focus; update server on pause
So the problem of fields losing focus every letter was fixed by moving
this.EContactInput into a class.

Then I just rode that excitement to updating the server which was pretty
easy.

Then I also added a timer that makes it so it doesn't update every
letter. NOTE: The timer doesn't work on remote debugging! Because the
two times are out of sync!